@@ -1,4 +1,4 @@
-class CreateListings < ActiveRecord::Migration
+class CreateListings < ActiveRecord::Migration[4.2]
   def up
     create_table :listings do |t|
       t.string   "name"
